Update components to 1.2.0-alpha1 version and create TextArea class
Our generator is not able to generate properly "value" accessors because "value" has "notify:true" and there is no anymore explicit "value-changed" event. So as a result synchronization, HasValue interface and setter correction are made here, in this patch in order to have the components behave properly.
This should be partially reverted once the issue #2100 is fixed.

Fix for #20